TechnologyManager finished so far. Worker, Building and Unit Interface still unclear, using current Interfaces. Additionally I added a string member variable and a getter function for that variable to the Entity class. This is necessary for the TechnologyManager to be able to determine which specific Entity he is dealing with. So to be able to use the getter method in Worker, Building and Unit-classes I added again the Inheritence from the Entity class for these three classes. Testing is still required, only succesfull compilation is granted. Unfortunately I won't be able to do much more today

+++ src/main.cpp
@@ -19,9 +19,9 @@ int main()
 		std::cout << tec->getSupplyCost() << "\t";
 		std::cout << tec->getBuildTime() << "\t";
 		std::vector<std::vector<std::shared_ptr<Technology>>> req = tec->getRequirements();
-		for(int i = 0; i < req.size(); ++i)
+		for(size_t i = 0; i < req.size(); ++i)
 		{
-			for (int j = 0; j < req[i].size(); ++j)
+			for (size_t j = 0; j < req[i].size(); ++j)
 				std::cout << req[i][j]->getName() <<"/"; 
 			std::cout << ",";
 		}
@@ -35,9 +35,9 @@ int main()
 		std::cout << tec->getSupplyCost() << "\t";
 		std::cout << tec->getBuildTime() << "\t";
 		std::vector<std::vector<std::shared_ptr<Technology>>> req = tec->getRequirements();
-		for(int i = 0; i < req.size(); ++i)
+		for(size_t i = 0; i < req.size(); ++i)
 		{
-			for (int j = 0; j < req[i].size(); ++j)
+			for (size_t j = 0; j < req[i].size(); ++j)
 				std::cout << req[i][j]->getName() <<"/";
 			std::cout << ",";
 		}
